backend/requirements.txt
fastapi
uvicorn
pymongo
sqlalchemy
psycopg2-binary
python-dotenv
